backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.146.0
google-generativeai==0.7.2
httpx==0.27.2
sqlalchemy==2.0.35
databases[sqlite]==0.9.0
python-dotenv==1.0.1
jinja2==3.1.4
